+++ code/campaign/src/main/java/com/adobe/marketing/mobile/campaign/LocalNotificationService.java
@@ -16,6 +16,8 @@ import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Build;
+
+import com.adobe.marketing.mobile.LocalNotificationHandler;
 import com.adobe.marketing.mobile.services.Log;
 import com.adobe.marketing.mobile.util.MapUtils;
 import java.security.SecureRandom;
